requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
h11==0.13.0
httpie==3.2.1
idna==3.4
multidict==6.0.2
numpy==1.23.2
outcome==1.2.0
pandas==1.4.3
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
schedule==1.1.0
selenium==4.4.3
six==1.16.0
slack-sdk==3.18.3
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
